# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Build with AK API & MCP Service",
    "version": "1.0.0",
    "description": "Official API and Model Context Protocol (MCP) server for authoring, validating, and submitting product showcases to the Build with AK Customer Directory on agentkit.best.",
    "contact": {
      "name": "AgentKit Support",
      "url": "https://agentkit.best",
      "email": "support@agentkit.best"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://bwak.agentkit.best",
      "description": "Production Cloudflare Worker Edge Server"
    }
  ],
  "security": [
    { "ApiKeyAuth": [] },
    { "OAuth2": ["build-with-ak:read"] }
  ],
  "paths": {
...
    "/api/build-with-ak/listing/analytics": {
      "get": {
        "summary": "Get owned product page analytics",
        "operationId": "getListingAnalytics",
        "servers": [
          {
            "url": "https://agentkit.best"
          },
          {
            "url": "https://staging.agentkit.best"
          }
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Owner-only analytics from first-party PostgreSQL. Optional listingId selects any owned historical listing; omitted selects the active non-rejected/non-archived listing. Inclusive UTC dates default to today and 29 days earlier. Maximum 366 days; to cannot be in the future. Unknown or duplicate parameters, invalid dates, and invalid ranges return 400. Authenticated browser sessions are also accepted by the upstream API. The MCP Worker exposes this through build_with_ak_get_analytics, not a REST proxy.",
        "parameters": [
          {
            "name": "listingId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Owned listing UUID; omitted selects the active listing."
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Inclusive UTC start, YYYY-MM-DD. Defaults to to minus 29 days."
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Inclusive UTC end, YYYY-MM-DD. Defaults to today; no future dates."
          }
        ],
        "responses": {
          "200": {
            "description": "Owned listing analytics with ascending, zero-filled daily counts.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingAnalytics"
                }
              }
            }
          },
          "400": {
            "description": "Invalid, unknown, or duplicate query parameter or invalid date range."
          },
          "401": {
            "description": "Authentication required or invalid API key."
          },
          "403": {
            "description": "Credential is not permitted to access this API."
          },
          "404": {
            "description": "No owned listing found; includes unknown or cross-owner listing IDs."
          },
          "429": {
            "description": "Rate limit exceeded."
          },
          "500": {
            "description": "Analytics retrieval failed."
          }
        }
      }
    },
...
    "/mcp": {
      "post": {
        "summary": "Streamable HTTP MCP Endpoint",
        "description": "Executes Model Context Protocol (MCP) JSON-RPC 2.0 commands over Streamable HTTP.",
        "operationId": "executeMcp",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "jsonrpc": { "type": "string", "example": "2.0" },
                  "id": { "oneOf": [{ "type": "string" }, { "type": "number" }] },
                  "method": { "type": "string", "example": "tools/call" },
                  "params": { "type": "object" }
                },
                "required": ["jsonrpc", "method"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Streamed MCP response or JSON-RPC result",
            "content": {
              "text/event-stream": {
                "schema": { "type": "string" }
              },
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "jsonrpc": { "type": "string", "example": "2.0" },
                    "id": { "oneOf": [{ "type": "string" }, { "type": "number" }] },
                    "result": { "type": "object" },
                    "error": { "type": "object" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — Missing or invalid API key or Bearer token with WWW-Authenticate challenge"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "AgentKit customer API key (ck_live_...)"
      },
      "OAuth2": {
        "type": "oauth2",
        "description": "OAuth 2.1 Authorization Code Flow with PKCE (S256)",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://agentkit.best/oauth/authorize",
            "tokenUrl": "https://agentkit.best/oauth/token",
            "scopes": {
              "build-with-ak:read": "Read product showcase draft and live listing",
              "build-with-ak:write": "Create, edit layout blocks, upload media, and submit showcase"
            }
          }
        }
      }
    },
    "schemas": {
      "AnalyticsMetrics": {
        "type": "object",
        "required": [
          "views",
          "outboundClicks",
          "referralConversions"
        ],
        "properties": {
          "views": {
            "type": "integer",
            "minimum": 0,
            "description": "Views deduplicated per listing, IP + user agent, and UTC day; not unique people across days."
          },
          "outboundClicks": {
            "type": "integer",
            "minimum": 0,
            "description": "Outbound redirect requests, not unique visitors."
          },
          "referralConversions": {
            "type": "null",
            "description": "Unavailable: conversion tracking is not instrumented. Not external product sales."
          }
        }
      },
      "ListingAnalytics": {
        "type": "object",
        "required": [
          "listing",
          "period",
          "source",
          "generatedAt",
          "totals",
          "daily"
        ],
        "properties": {
          "listing": {
            "type": "object",
            "required": [
              "id",
              "slug"
            ],
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "slug": {
                "type": "string"
              }
            }
          },
          "period": {
            "type": "object",
            "required": [
              "from",
              "to",
              "timezone"
            ],
            "properties": {
              "from": {
                "type": "string",
                "format": "date"
              },
              "to": {
                "type": "string",
                "format": "date"
              },
              "timezone": {
                "type": "string",
                "const": "UTC"
              }
            }
          },
          "source": {
            "type": "string",
            "const": "first_party_postgresql"
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "totals": {
            "$ref": "#/components/schemas/AnalyticsMetrics"
          },
          "daily": {
            "type": "array",
            "minItems": 1,
            "maxItems": 366,
            "description": "Ascending UTC dates, including zero-count view/click days. Referral conversions are null on every day because tracking is unavailable.",
            "items": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/AnalyticsMetrics"
                },
                {
                  "type": "object",
                  "required": [
                    "date"
                  ],
                  "properties": {
                    "date": {
                      "type": "string",
                      "format": "date"
                    }
                  }
                }
              ]
            }
          }
        }
      },
...
    }
  }
}